requirements.txt
numpy==1.19.5
Pillow==8.2.0
requests==2.25.1
urllib3==1.26.4
